+++ src/payloads/hello_world.cpp
@@ -6,7 +6,10 @@ int main(int argc, char** argv)
 {
     bmboot::notifyPayloadStarted();
 
-    printf("Hello world from CPU%d! Our base address is %p.\n", bmboot::getCpuIndex(), &_vector_table);
+    printf("Hello world from CPU%d! Our base address is %p and the Payload Argument is %ld.\n",
+           bmboot::getCpuIndex(),
+           (void*) &_vector_table,
+           bmboot::getPayloadArgument());
 
     for (;;) {}
 }